src/requirements.txt
python-dotenv~=0.21.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-zipkin==1.22.0
psutil==5.8.0
paho-mqtt==1.6.1
simple-profile==1.1.2
xycrypto==0.9